requirements.txt
ovh==1.1.0

python-dotenv
requests~=2.31.0
